+++ api/V0ComposeKillService.go
@@ -3,7 +3,6 @@ package api
 import (
 	"net/http"
 
-	mesosutil "github.com/AVENTER-UG/mesos-util"
 	"github.com/gorilla/mux"
 	"github.com/sirupsen/logrus"
 )
@@ -37,7 +36,7 @@ func (e *API) V0ComposeKillService(w http.ResponseWriter, r *http.Request) {
 	for keys.Next(e.Redis.CTX) {
 		key := e.Redis.GetRedisKey(keys.Val())
 
-		task := mesosutil.DecodeTask(key)
+		task := e.Mesos.DecodeTask(key)
 		task.State = "__KILL"
 		e.Redis.SaveTaskRedis(task)