+++ api/V0ShowAllTasks.go
@@ -4,8 +4,8 @@ import (
 	"encoding/json"
 	"net/http"
 
-	mesosutil "github.com/AVENTER-UG/mesos-util"
-	mesosproto "github.com/AVENTER-UG/mesos-util/proto"
+	mesosproto "github.com/AVENTER-UG/mesos-compose/proto"
+	cfg "github.com/AVENTER-UG/mesos-compose/types"
 	util "github.com/AVENTER-UG/util/util"
 	"github.com/sirupsen/logrus"
 )
@@ -27,7 +27,7 @@ func (e *API) V0ShowAllTasks(w http.ResponseWriter, r *http.Request) {
 
 	keys := e.Redis.GetAllRedisKeys(e.Framework.FrameworkName + ":*")
 
-	var list []mesosutil.Command
+	var list []cfg.Command
 
 	for keys.Next(e.Redis.CTX) {
 		// ignore redis keys if they are not mesos tasks
@@ -36,7 +36,7 @@ func (e *API) V0ShowAllTasks(w http.ResponseWriter, r *http.Request) {
 		}
 
 		key := e.Redis.GetRedisKey(keys.Val())
-		task := mesosutil.DecodeTask(key)
+		task := e.Mesos.DecodeTask(key)
 
 		task.Environment = mesosproto.Environment{}